Code Refactoring for Beneficiary DTOs.
