+++ src/main/java/com/iemr/hwc/fhir/utils/mapper/MapperUtils.java
@@ -1,7 +1,7 @@
 package com.iemr.hwc.fhir.utils.mapper;
 
-import com.iemr.hwc.fhir.dto.benDemographics.DemographicsDTO;
-import com.iemr.hwc.fhir.dto.benPhone.PhoneDetailsDTO;
+import com.iemr.hwc.fhir.dto.beneficiary.benDemographics.DemographicsDTO;
+import com.iemr.hwc.fhir.dto.beneficiary.benPhone.PhoneDetailsDTO;
 import com.iemr.hwc.fhir.dto.beneficiary.BeneficiaryDTO;
 import com.iemr.hwc.fhir.model.patient.PatientExt;
 import org.mapstruct.Mapper;
